#344: Deserialization produces graph nodes
