+++ src/test/resources/com/atomist/tree/marshal/withLinks.json
@@ -1,0 +1,105 @@
+[
+  {
+    "build_url":"https://travis-ci.org/something",
+    "id":"192756197",
+    "timestamp":"2017-01-17T17:21:48.374Z",
+    "started_at":"2017-01-17T17:18:57Z",
+    "nodeId":4770,
+    "name":"333",
+    "compare_url":"https://github.com/something",
+    "status":"Passed",
+    "type":[
+      "Build"
+    ],
+    "finished_at":"2017-01-17T17:21:45Z"
+  },
+  {
+    "message":"Commit message",
+    "sha":"sha",
+    "timestamp":"2017-01-17T11:18:51-06:00",
+    "nodeId":4767,
+    "type":[
+      "Commit"
+    ]
+  },
+  {
+    "name":"user-name",
+    "login":"user-login",
+    "email":"user-email",
+    "nodeId":2900,
+    "type":[
+      "GitHubId"
+    ]
+  },
+  {
+    "owner":"owner",
+    "name":"repo-name",
+    "nodeId":2391,
+    "type":[
+      "Repo"
+    ]
+  },
+  {
+    "name":"channel-name",
+    "id":"channel-id",
+    "nodeId":2200,
+    "type":[
+      "ChatChannel",
+      "SlackChannel"
+    ]
+  },
+  {
+    "before":"before-sha",
+    "after":"after-sha",
+    "branch":"git-branch",
+    "timestamp":"2017-01-17T17:18:52.943Z",
+    "nodeId":4766,
+    "type":[
+      "Push", "-dynamic"
+    ]
+  },
+  {
+    "startNodeId":4767,
+    "endNodeId":4770,
+    "type":"HAS_BUILD"
+  },
+  {
+    "startNodeId":4767,
+    "endNodeId":2900,
+    "type":"AUTHOR"
+  },
+  {
+    "startNodeId":4770,
+    "endNodeId":2391,
+    "type":"ON",
+    "cardinality": "1:1",
+    "random-nonsense": "savage,ruthless"
+  },
+  {
+    "startNodeId":4770,
+    "endNodeId":2391,
+    "type":"ONM",
+    "cardinality": "1:M"
+  },
+  {
+    "startNodeId":2391,
+    "endNodeId":2200,
+    "type":"CHANNEL"
+  },
+  {
+    "startNodeId":4770,
+    "endNodeId":4766,
+    "type":"TRIGGERED_BY"
+  },
+  {
+    "startNodeId":4766,
+    "endNodeId":4767,
+    "type":"CONTAINS",
+    "cardinality": "1:M"
+  },
+  {
+    "startNodeId":4767,
+    "endNodeId":2900,
+    "type":"AUTHOR"
+  }
+]
